Moved ServerResponseData to wrapper directory.
