+++ android/app/src/main/java/com/koleff/kare_android/data/model/wrapper/GetExercisesWrapper.kt
@@ -1,7 +1,6 @@
 package com.koleff.kare_android.data.model.wrapper
 
 import com.koleff.kare_android.data.model.response.GetExercisesResponse
-import com.koleff.kare_android.data.model.response.base_response.ServerResponseData
 
 class GetExercisesWrapper(getExercisesResponse: GetExercisesResponse) :
     ServerResponseData(getExercisesResponse) {
