+++ android/app/src/main/java/com/koleff/kare_android/data/model/wrapper/ServerResponseData.kt
@@ -1,4 +1,7 @@
-package com.koleff.kare_android.data.model.response.base_response
+package com.koleff.kare_android.data.model.wrapper
+
+import com.koleff.kare_android.data.model.response.base_response.BaseResponse
+import com.koleff.kare_android.data.model.response.base_response.KareError
 
 open class ServerResponseData(
     var isSuccessful: Boolean,
